Add Mahalanobis detector
