+++ alibi_detect/od/__init__.py
@@ -1,7 +1,7 @@
 from alibi_detect.utils.missing_optional_dependency import import_optional
 
 from .isolationforest import IForest
-from .mahalanobis import Mahalanobis
+from .stateful_mahalanobis import Mahalanobis
 from .sr import SpectralResidual
 
 from alibi_detect.od.base import TransformProtocol, transform_protocols
